# perf: make TCP_NODELAY configurable to improve query performance