--- a/src/clickhouse-cpp/clickhouse/client.cpp
+++ b/src/clickhouse-cpp/clickhouse/client.cpp
@@ -356,6 +356,10 @@ void Client::Impl::ResetConnection() {
                           options_.tcp_keepalive_cnt);
     }
 
+    if (options_.tcp_nodelay) {
+        s.SetTcpNoDelay(options_.tcp_nodelay);
+    }
+
     socket_ = std::move(s);
     socket_input_ = SocketInput(socket_);
     socket_output_ = SocketOutput(socket_);