@jerrywuz 2025-07-20
To verify the performance improvement, we can examine the insertion timestamps by executing the following SQL in ClickHouse. This reveals noticeable differences:
```sql
SELECT
    event_time,
    query,
    query_duration_ms,
    type
FROM system.query_log
WHERE
    query_kind = 'Insert'
ORDER BY event_time DESC
LIMIT 20;
```